Food Management System
This is a food management system where user can access the items on the
menu and also check the crucial details such as the content and
materials in the food such as fats and sugars. This also includes a
recommender system that recommends users with some food items based on
their interests for healthy diet.